E.Template

Description

A text file with sensitive data.

Tem could contain plain text, plain values PlainValue that would be encrypted, and references to already encrypted values ValRef.

PlainValue syntax

{{P|<variable name>|<encryption algorithm name>|<arguments>|<content>}}

E.g. {{P|username|gpgme|keyId = foobar|dmalikov}}.

ValRef syntax

{{E|<variable name>}}

E.g. {{E|username}}.

<variable name>. Name of the variable stored in Metadata which could be referenced in ValRef.

<encryption algorithm name>. AlgName provided by one of the algorithms Algs used for encryption/decryption.

<arguments>. Arguments that encryption algorithm Algs use during encryption/decryption.

<content>. Sensitive data to encrypt.

Template

data Tem

Representation of encrypted and/or to-be-encrypted content.

Constructors

Nil
End of Tem.

!Text `Txt` !Tem
Chunk with text.

!PlainValue `Val` !Tem
Chunk with plain value (intended to be encrypted).

!ValRef `Ref` !Tem
Chunk with reference to already encrypted value.

txt :: Text -> Tem

Convert Text to Tem.

val :: PlainValue -> Tem

Convert PlainValue to Tem.

ref :: ValRef -> Tem

Convert ValRef to Tem.

Value

data PlainValue

Representation of plain value.

Constructors

PlainValue !ValName !AlgName !Args !PlainContent

data PlainContent

Plain text.

Constructors

PlainContent

Fields

unPlainContent :: !Text

data EncValue

Representation of encrypted value (in Metadata).

Constructors

EncValue !AlgName !Args !EncContent

data EncContent

Ciphered text.

Constructors

EncContent

Fields

unEncContent :: !Text

newtype ValName

Value name.

Constructors

ValName

Fields

unValName :: Text

data ValRef

Reference to an encrypted value.

Constructors

ValRef

Fields

unValRef :: !ValName

Encryption/decryption arguments

data Args

Bunch of arguments that could be useful for encryption/decryption.

Constructors

Args

Fields

unArgs :: HashMap ArgName ArgValue

data ArgName

Typed key of Args hashmap.

Constructors

ArgName

Fields

unArgName :: !Text

data ArgValue

Typed value of Args hashmap.

Constructors

ArgValue

Fields

unArgValue :: !Text

newtype AlgName

Encryption/decryption algorithm name. E.g. gpg.

Constructors

AlgName

Fields

unAlgName :: Text

Args operations

arg :: ArgName -> ArgValue -> Args

Args constructor.

lookupArg :: Text -> Args -> Maybe ArgValue

Lookup arg.

Operations on Tem

normalize :: Tem -> Tem

Normalize Tem squashing all neighboring Txt chunks.

Parsers

parseMaybe :: Parser a -> Text -> Maybe a

Apply parser and maybe return result.

parseArgNV :: Parser (ArgName, ArgValue)

Parser for key-value pair (ArgName, ArgValue).

parseArgs :: Parser Args

Parser for Args.

parsePlainValue :: Parser PlainValue

Parser for PlainValue.

parseValRef :: Parser ValRef

Parser for ValRef.

parseTem :: Parser Tem

Parser for Tem.

Serializing routines

class Serialize a where

Bare minimum of serializing to and from Text with no information of deserializing failure reason.

Minimal complete definition

encode, decode

Methods

encode :: a -> Text

decode :: Text -> Maybe a