System.Template.Encrypt

Description

FilePath templating mechanism with a power of Encrypt.

Documentation

encrypt

Arguments

:: FilePath
Input file.

-> FilePath
Output file.

-> IO (Maybe EncryptSystemError)

For a given template update all the plain-text holes with encrypted values and produce a file in a given filepath.

decrypt

Arguments

:: FilePath
Input file.

-> FilePath
Output file.

-> IO (Maybe DecryptSystemError)

For a given template update all the encrypted-text holes with decrypted values and produce a file in a given filepath.

data EncryptSystemError

Templating error occurred during encryption.

Constructors

EncryptFSError FSError
FSError occurred during encryption.

EncryptTemplateError EncryptTemplateError
Internal EncryptTemplateError.

data DecryptSystemError

Templating error occurred during decryption.

Constructors

DecryptFSError FSError
FSError occurred during decryption.

DecryptTemplateError DecryptTemplateError
Internal DecryptTemplateError.

data FSError

File system related error.

Constructors

InputFileNotFound
Input file doesn't exist.

OutputDirectoryNotFound
Directory where output should be generated at doesn't exist.